feat(service): make all service methods optional for cleaner implementations

+++ projects/ngrx-auto-entity/src/lib/service.ts
@@ -35,21 +35,21 @@ export interface IEntityWithRangeInfo<TModel> {
 }
 
 export interface IAutoEntityService<TModel> {
-  load(entityInfo: IEntityInfo, keys: any, relationKeys?: any): Observable<TModel>;
+  load?(entityInfo: IEntityInfo, keys: any, relationKeys?: any): Observable<TModel>;
 
-  loadAll(entityInfo: IEntityInfo, relationKeys?: any): Observable<TModel[]>;
+  loadAll?(entityInfo: IEntityInfo, relationKeys?: any): Observable<TModel[]>;
 
-  loadPage(entityInfo: IEntityInfo, page: Page, relationKeys?: any): Observable<IEntityWithPageInfo<TModel>>;
+  loadPage?(entityInfo: IEntityInfo, page: Page, relationKeys?: any): Observable<IEntityWithPageInfo<TModel>>;
 
-  loadRange(entityInfo: IEntityInfo, range: Range, relationKeys?: any): Observable<IEntityWithRangeInfo<TModel>>;
+  loadRange?(entityInfo: IEntityInfo, range: Range, relationKeys?: any): Observable<IEntityWithRangeInfo<TModel>>;
 
-  create(entityInfo: IEntityInfo, entity: TModel, relationKeys?: any): Observable<TModel>;
+  create?(entityInfo: IEntityInfo, entity: TModel, relationKeys?: any): Observable<TModel>;
 
-  update(entityInfo: IEntityInfo, entity: TModel, relationKeys?: any): Observable<TModel>;
+  update?(entityInfo: IEntityInfo, entity: TModel, relationKeys?: any): Observable<TModel>;
 
-  replace(entityInfo: IEntityInfo, entity: TModel, relationKeys?: any): Observable<TModel>;
+  replace?(entityInfo: IEntityInfo, entity: TModel, relationKeys?: any): Observable<TModel>;
 
-  delete(entityInfo: IEntityInfo, entity: TModel, relationKeys?: any): Observable<TModel>;
+  delete?(entityInfo: IEntityInfo, entity: TModel, relationKeys?: any): Observable<TModel>;
 }
 
 /**
@@ -63,6 +63,14 @@ export class NgrxAutoEntityService {
   load<TModel>(entityInfo: IEntityInfo, keys: any, relationKeys?: any): Observable<IEntityRef<TModel>> {
     try {
       const service = this.getService<TModel>(entityInfo);
+
+      if (!service.load) {
+        return throwError({ info: entityInfo, message: 'Entity service "load" is not implemented' });
+      }
+
+      if (typeof service.load !== 'function') {
+        return throwError({ info: entityInfo, message: 'Entity service "load" is not a function' });
+      }
 
       return service.load(entityInfo, keys, relationKeys).pipe(
         map(entity => ({
@@ -87,6 +95,15 @@ export class NgrxAutoEntityService {
   loadAll<TModel>(entityInfo: IEntityInfo, relationKeys?: any): Observable<IEntityRef<TModel[]>> {
     try {
       const service = this.getService<TModel>(entityInfo);
+
+      if (!service.loadAll) {
+        return throwError({ info: entityInfo, message: 'Entity service "loadAll" is not implemented' });
+      }
+
+      if (typeof service.loadAll !== 'function') {
+        return throwError({ info: entityInfo, message: 'Entity service "loadAll" is not a function' });
+      }
+
       return service.loadAll(entityInfo, relationKeys).pipe(
         map((entities: TModel[]) => ({
           info: entityInfo,
@@ -110,6 +127,15 @@ export class NgrxAutoEntityService {
   loadPage<TModel>(entityInfo: IEntityInfo, page: Page, relationKeys?: any): Observable<IEntityPageRef<TModel>> {
     try {
       const service = this.getService<TModel>(entityInfo);
+
+      if (!service.loadPage) {
+        return throwError({ info: entityInfo, message: 'Entity service "loadPage" is not implemented' });
+      }
+
+      if (typeof service.loadPage !== 'function') {
+        return throwError({ info: entityInfo, message: 'Entity service "loadPage" is not a function' });
+      }
+
       return service.loadPage(entityInfo, page, relationKeys).pipe(
         map((result: IEntityWithPageInfo<TModel>) => ({
           info: entityInfo,
@@ -134,6 +160,15 @@ export class NgrxAutoEntityService {
   loadRange<TModel>(entityInfo: IEntityInfo, range: Range, relationKeys?: any): Observable<IEntityRangeRef<TModel>> {
     try {
       const service = this.getService<TModel>(entityInfo);
+
+      if (!service.loadRange) {
+        return throwError({ info: entityInfo, message: 'Entity service "loadRange" is not implemented' });
+      }
+
+      if (typeof service.loadRange !== 'function') {
+        return throwError({ info: entityInfo, message: 'Entity service "loadRange" is not a function' });
+      }
+
       return service.loadRange(entityInfo, range, relationKeys).pipe(
         map((result: IEntityWithRangeInfo<TModel>) => ({
           info: entityInfo,
@@ -159,6 +194,14 @@ export class NgrxAutoEntityService {
     try {
       const service = this.getService<TModel>(entityInfo);
 
+      if (!service.create) {
+        return throwError({ info: entityInfo, message: 'Entity service "create" is not implemented' });
+      }
+
+      if (typeof service.update !== 'function') {
+        return throwError({ info: entityInfo, message: 'Entity service "update" is not a function' });
+      }
+
       return service.create(entityInfo, entity, relationKeys).pipe(
         map(savedEntity => ({
           info: entityInfo,
@@ -182,6 +225,14 @@ export class NgrxAutoEntityService {
   update<TModel>(entityInfo: IEntityInfo, entity: TModel, relationKeys?: any): Observable<IEntityRef<TModel>> {
     try {
       const service = this.getService<TModel>(entityInfo);
+
+      if (!service.update) {
+        return throwError({ info: entityInfo, message: 'Entity service "update" is not implemented' });
+      }
+
+      if (typeof service.update !== 'function') {
+        return throwError({ info: entityInfo, message: 'Entity service "update" is not a function' });
+      }
 
       return service.update(entityInfo, entity, relationKeys).pipe(
         map(savedEntity => ({
@@ -207,6 +258,14 @@ export class NgrxAutoEntityService {
     try {
       const service = this.getService<TModel>(entityInfo);
 
+      if (!service.replace) {
+        return throwError({ info: entityInfo, message: 'Entity service "replace" is not implemented' });
+      }
+
+      if (typeof service.replace !== 'function') {
+        return throwError({ info: entityInfo, message: 'Entity service "replace" is not a function' });
+      }
+
       return service.replace(entityInfo, entity, relationKeys).pipe(
         map(savedEntity => ({
           info: entityInfo,
@@ -230,6 +289,14 @@ export class NgrxAutoEntityService {
   delete<TModel>(entityInfo: IEntityInfo, entity: TModel, relationKeys?: any): Observable<IEntityRef<TModel>> {
     try {
       const service = this.getService<TModel>(entityInfo);
+
+      if (!service.delete) {
+        return throwError({ info: entityInfo, message: 'Entity service "delete" is not implemented' });
+      }
+
+      if (typeof service.delete !== 'function') {
+        return throwError({ info: entityInfo, message: 'Entity service "delete" is not a function' });
+      }
 
       return service.delete(entityInfo, entity, relationKeys).pipe(
         map(deletedEntity => ({
